refactor: messenger 패키지로 변경 및 MessengerClient 인터페이스 적용 (#172)
