+++ src/main/kotlin/co/yappuworld/external/messenger/DiscordProperties.kt
@@ -1,4 +1,4 @@
-package co.yappuworld.external.discord
+package co.yappuworld.external.messenger
 
 import org.springframework.boot.context.properties.ConfigurationProperties
