+++ src/main/kotlin/co/yappuworld/external/messenger/MessengerClient.kt
@@ -1,0 +1,8 @@
+package co.yappuworld.external.messenger
+
+import co.yappuworld.external.messenger.dto.DiscordEmbed
+
+interface MessengerClient {
+    fun send(content: String)
+    fun send(embed: DiscordEmbed)
+}
