+++ src/main/kotlin/co/yappuworld/user/infrastructure/DiscordUserSystemNotifier.kt
@@ -1,8 +1,8 @@
 package co.yappuworld.user.infrastructure
 
-import co.yappuworld.external.discord.DiscordClient
-import co.yappuworld.external.discord.dto.DiscordEmbed
-import co.yappuworld.external.discord.dto.DiscordEmbedField
+import co.yappuworld.external.messenger.MessengerClient
+import co.yappuworld.external.messenger.dto.DiscordEmbed
+import co.yappuworld.external.messenger.dto.DiscordEmbedField
 import co.yappuworld.global.property.AdminProperties
 import co.yappuworld.user.infrastructure.entity.SignUpApplicationEntity
 import org.springframework.context.annotation.Primary
@@ -11,12 +11,12 @@ import org.springframework.stereotype.Component
 @Component
 @Primary
 class DiscordUserSystemNotifier(
-    private val discordClient: DiscordClient,
+    private val messengerClient: MessengerClient,
     private val adminProperties: AdminProperties
 ) : UserSystemNotifier {
 
     override fun notifySignUpRequestReceived(signUpApplication: SignUpApplicationEntity) {
-        discordClient.send(
+        messengerClient.send(
             DiscordEmbed.info(
                 title = "💡 회원가입 신청을 확인해주세요 💡",
                 url = "${adminProperties.domain}/admin/members/application",
