+++ src/test/kotlin/co/yappuworld/external/messenger/DiscordClientTest.kt
@@ -1,4 +1,4 @@
-package co.yappuworld.external.discord
+package co.yappuworld.external.messenger
 
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
@@ -6,11 +6,11 @@ import org.springframework.boot.test.context.SpringBootTest
 
 @SpringBootTest
 class DiscordClientTest @Autowired constructor(
-    private val discordClient: DiscordClient
+    private val messengerClient: MessengerClient
 ) {
 
     @Test
     fun test() {
-        discordClient.send("테스트 컨텐츠")
+        messengerClient.send("테스트 컨텐츠")
     }
 }
